Adds OCCI support.
Implements blueprint bexar-open-cloud-compute-interface.

This implements an OpenStack service that runs out of nova-api.
It is implemented using both the OpenStack service and WSGI frameworks. On
start it will serve its functionality over HTTP on port 8787 as described
in the OCCI specifications. It is compliant as per the set of OCCI
specifications (GFD.183, GFD.184 and GFD.185) and implements all mandatory
features. It also leverages the OCCI core model to expose OpenStack-specific
features in an OCCI fashion. There is further detail on its core and OpenStack
specific usage at http://wiki.openstack.org/occi. The implementation is
co-funded by Intel Labs Europe Cloud Services Lab and FI-ware.

Signed-off-by: Andy Edmonds <andy@edmonds.be>
Change-Id: I402cdee80b42c8b12f4c6e954187fa544f959b4f